Merged PR 6456: Added check for whitelisting scripts
Changes are based upon the pentest findings in items: _RED-14-008356-001: Deployment scripts allow insecure_ configuration.

We have created a new piece of functionality to make sure that allow 'any' (`0.0.0.0/0`) access is not allowed.  This, because you should be able to delete the above, but not be able to add it.

**Changes**
- Added extra check to the function 'Assert-CIDR'.

**Resources**
Added the check to the following resources (other resources already had this check):

- AppService
- Container registry
- FunctionApp
- Keyvault
- MySQL
- PostgreSQL
- RedisCache
- ServiceBus
- SQL Server
- Storage Account

Testen staan in de gelinkte story.

Related work items: #18652, #18655