Merged PR 6543: PurgeProtection for keyvault
**Swap App Service deprecated**
- Added deprecation flag to script `Swap-App-Service.ps1`.

**Enabled purge protection for keyvault**
- Added parameter `KeyvaultSku` and `KeyvaultPurgeProtectionEnabled` to `Create-Keyvault.ps1`.
- Added `-ForceDisablePurgeProtection` to avoid creating keyvaults without purge-protection enabled unintentionally
- Recovering soft-deleted keyvault if it already exists with same name.

**Application Gateway Entrypoints**

During the entrypoint creation, we now check if the certificate exists in soft-deleted state. If so, we recover the certificate (and then check if we need to update it). If not, we continue the flow as before.

During the deprovisioning of the entrypoint, we now only soft-delete the certificate and don't purge it afterwards.

Tests can be found in attached workitem:  #18845.

Related work items: #18845, #18846