Merged PR 6965: Added Servicebus zone redundancy
Added the possibility to enable zone redundancy for servicebus. This cannot be done with the Azure CLI or with the powershell module, so az rest was used for this specific usecase.

Tests have been done with @<Noëlle Heerink - Wijnja> and the results have been validated by her.

Related work items: #20005