Merged PR 10401: priority parameter optional when creating new application gateway rules
Priority parameter optional when creating new application gateway rules.

The previous fix where we query the application gateway for the last priority results in failures when deploying to the production gateway, where the priority query retuns `null`.

The production gateway does not seem to require that we pass the priority explicitly (possibly due to some A-B deployment strategy from microsoft?) so i've altered the query and added a null-check.