Merged PR 15127: Prevent depleting of agents for regressionpipeline
We see that when the regression pipeline runs it takes up all resources (Agents) for deploying. The regression pipeline now also runs when there is a change in the azdocsv2 (for which no regressionset is defined).
Option: limit the run hours of this pipeline outside office-hours by using a new environment with this check.