Merged PR 40923: feat: Set defaultBlobAccessTier based on storageAccountSku
feat: Set defaultBlobAccessTier based on storageAccountSku

The code changes in `storageAccounts.bicep` modify the `accessTier` property of the `storageAccount` resource. Instead of using the `defaultBlobAccessTier` directly, the `effectiveBlobAccessTier` variable is introduced to set the `accessTier` based on the value of `storageAccountSku`. If `storageAccountSku` contains the word "Premium", `effectiveBlobAccessTier` is set to null; otherwise, it is set to `defaultBlobAccessTier`. This change improves the flexibility of setting the `accessTier` for the storage account.